+++ src/AppRouter.tsx
@@ -50,7 +50,6 @@ export default class AppRouter extends React.Component<AppRouterProps, AppRouter
     ErrorComponent: ({ err }: { err: string | Error}) => <div>{ typeof err === 'string' ? err : err?.message }</div>,
     LoadingComponent: <div>Loading...</div>,
     NotFoundComponent: <div>NotFound</div>,
-    shouldAssetsRemove: () => true,
     onAppEnter: () => {},
     onAppLeave: () => {},
     onLoadingApp: () => {},
@@ -93,7 +92,6 @@ export default class AppRouter extends React.Component<AppRouterProps, AppRouter
      */
     const { shouldAssetsRemove, onAppEnter, onAppLeave, fetch, basename } = this.props;
     start({
-      shouldAssetsRemove,
       onAppLeave,
       onAppEnter,
       onLoadingApp: this.loadingApp,
@@ -102,7 +100,9 @@ export default class AppRouter extends React.Component<AppRouterProps, AppRouter
       reroute: this.handleRouteChange,
       fetch,
       basename,
+      ...(shouldAssetsRemove ? { shouldAssetsRemove } : {}),
     });
+
     this.setState({ started: true });
   }