+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 See [https://github.com/ice-lab/icestark/releases](https://github.com/ice-lab/icestark/releases) for what has changed in each version of icestark.
 
+## 2.7.0
+
+- [feat] cache css by default. ([#373](https://github.com/ice-lab/icestark/issues/373))
+- [feat] appHistory and <Link /> can both take state. ([#477](https://github.com/ice-lab/icestark/issues/477))
+- [chore] narrow scope of `import`'s error. ([#466](https://github.com/ice-lab/icestark/issues/466))
+- [chore] add missing props for lifecycles. ([#440](https://github.com/ice-lab/icestark/issues/440))
+
 ## 2.6.2
 
 - [fix] avoid to append duplicate assets. ([#331](https://github.com/ice-lab/icestark/issues/331))